+++ pkgbuilder/__main__.py
@@ -17,8 +17,9 @@ def write_yaml_file(yaml_file_path: str, obj):
 class Pkgbuilder:
     def __init__(self, config: dict):
         self.config = config
+        self.output = ""
 
-    def print(self):
+    def configurate(self):
         for person in self.config["people"]:
             self.print_person(person)
         self.print_header()
@@ -34,28 +35,28 @@ class Pkgbuilder:
                 opts.append(k)
             else:
                 opts.append("!" + k)
-        print("options=(%s)" % " ".join(opts))
+        self.output += ("options=(%s)\n" % " ".join(opts))
 
     def print_person(self, person: dict):
-        print("# %s: %s %s <%s>" % (
+        self.output += ("# %s: %s %s <%s>\n" % (
             person["role"], person["surname"], person["name"], person["email"]
         ))
 
     def print_header(self):
-        print("pkgname=%s" % self.config["pkgname"])
-        print("pkgver=%s" % self.config["pkgver"])
-        print("pkgrel=%s" % self.config["pkgrel"])
-        print("pkgdesc=\'%s\'" % self.config["pkgdesc"])
-        print("license=('%s')" % self.config["license"])
-        print("arch=(%s)" % " ".join(["'%s'" % arch for arch in self.config["arch"]]))
-        print("depends=(%s)" % " ".join(["'%s'" % depends for depends in self.config["depends"]]))
-        print("makedepends=(%s)" % " ".join(["'%s'" % makedepends for makedepends in self.config["makedepends"]]))
-        print("url='%s'" % self.config["url"])
+        self.output += ("pkgname=%s\n" % self.config["pkgname"])
+        self.output += ("pkgver=%s\n" % self.config["pkgver"])
+        self.output += ("pkgrel=%s\n" % self.config["pkgrel"])
+        self.output += ("pkgdesc=\'%s\'\n" % self.config["pkgdesc"])
+        self.output += ("license=('%s')\n" % self.config["license"])
+        self.output += ("arch=(%s)\n" % " ".join(["'%s'" % arch for arch in self.config["arch"]]))
+        self.output += ("depends=(%s)\n" % " ".join(["'%s'" % depends for depends in self.config["depends"]]))
+        self.output += ("makedepends=(%s)\n" % " ".join(["'%s'" % makedepends for makedepends in self.config["makedepends"]]))
+        self.output += ("url='%s'\n" % self.config["url"])
 
     def print_sources(self):
-        print("_archive='%s'" % self.config["archive"])
-        print("source=(\"$pkgname-$pkgver.tar.gz::$_archive/$pkgver.tar.gz\")")
-        print("sha256sums=('%s')" % self.compute_hash())
+        self.output += ("_archive='%s'\n" % self.config["archive"])
+        self.output += ("source=(\"$pkgname-$pkgver.tar.gz::$_archive/$pkgver.tar.gz\")\n")
+        self.output += ("sha256sums=('%s')\n" % self.compute_hash())
 
     def compute_hash(self):
         link = "%s/%s.tar.gz" % (self.config["archive"], self.config["pkgver"])
@@ -68,30 +69,35 @@ class Pkgbuilder:
         return hash.hex()
 
     def print_build(self):
-        print("build() {")
-        print("  cd %s" % "$pkgname-$pkgver")
+        self.output += ("build() {\n")
+        self.output += ("  cd %s\n" % "$pkgname-$pkgver")
         for line in self.config["build"]:
-            print("  " + line)
-        print("}")
+            self.output += ("  " + line + "\n")
+        self.output += ("}\n")
 
     def print_package(self):
-        print("package() {")
-        print("  cd %s" % "$pkgname-$pkgver")
+        self.output += ("package() {\n")
+        self.output += ("  cd %s\n" % "$pkgname-$pkgver")
         for line in self.config["package"]:
-            print("  " + line)
-        print("}")
+            self.output += ("  " + line + "\n")
+        self.output += ("}\n")
 
 def main_cli():
     cli = argparse.ArgumentParser(description="Declarative PKGBUILD automatic generation")
     cli.add_argument("-V", "--version", type=str, default=None, help="override Version field from CLI")
     cli.add_argument("-v", "--verbose", action='store_true', default=False, help="Enables verbose log")
+    cli.add_argument("-o", "--output", default="PKGBUILD", help="Output path")
     will = cli.parse_args(sys.argv[1:])
     if will.verbose:
         logging.getLogger().setLevel(logging.DEBUG)
     config = read_yaml_file('pkgbuild.yaml')
     if will.version is not None:
         config['pkgver'] = will.version
-    Pkgbuilder(config).print()
+    pkgbuild = Pkgbuilder(config)
+    pkgbuild.configurate()
+
+    with open(will.output, mode="w", encoding="utf-8") as out:
+        out.write(pkgbuild.output)
 
 if __name__ == "__main__":
     main_cli()